requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colour-science==0.4.3
contourpy==1.1.0
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.12.2
fonttools==4.42.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
kornia==0.7.2
kornia_rs==0.1.2
lazy_loader==0.3
lightning==2.2.0.post0
lightning-utilities==0.10.1
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
pandas==2.2.1
Pillow==10.0.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==2.2.0.post0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.4.1.post1
scipy==1.11.2
sentry-sdk==1.40.5
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.3.0
tifffile==2023.8.12
torch==2.0.1
torchmetrics==1.3.1
torchvision==0.15.2
tqdm==4.66.2
typing_extensions==4.7.1
tzdata==2024.1
urllib3==2.0.4
wandb==0.16.3
yarl==1.9.4
